requirements.txt
attrs==25.1.0
beautifulsoup4==4.13.3
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
h11==0.14.0
idna==3.10
numpy==2.2.2
outcome==1.3.0.post0
pandas==2.2.3
pathvalidate==3.2.3
psutil==6.1.1
pycparser==2.22
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
requests==2.32.3
selenium==4.28.1
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
trio==0.28.0
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2025.1
undetected-chromedriver==3.5.5
urllib3==2.3.0
websocket-client==1.8.0
websockets==14.2
wsproto==1.2.0
